added issue logic in the controller
